Rename LocalDateTime to local::DateTime
This has been changed to mimic things like io::Error and io::Result, and fmt::Result, among others. We don't need to be so specific with our names when the language has features to sort them out for you!

I've picked "local::DateTime" for code, and "local::DateTime as LocalDateTime" for some of the tests, because I couldn't fully make my mind up about which to use. We'll see how it goes.