Pass around Year values more often
This commit changes the year() DatePiece accessor to return a Year instead of an i64, and changes the Date constructors to accept either a Year or an i64.

I mean, what good are having all these type-safety values if we're not going to use them? Other than to fix #12?

There are also a lot more 'translation' functions for Year, such as Deref and From, to help people use this type.