Replace range check 'if's with 'try!'s
This is paving the way for extra information in the OutOfRange struct (the field, the range, and the value it tried to set).

Instead of a blanket OutOfRange being returned, now there's exactly one early return for each field that's being checked.

`days_since_epoch` has been changed to not do any checking itself, since there are a few places where the result was just being unwrapped anyway. It makes more sense for every other use to do the checking itself beforehand.